Remove tint.nvim
tint.nvim is a plugin that provides a terminal emulator in a vim buffer.

I have been using it for a while, but I have found that I don't use it
enough to justify the extra complexity it adds to my setup.

I have been using the built-in terminal emulator for a while now and I
haven't missed tint.nvim.